+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Feature
 - Make Mosmix and DMO a lot faster for multiple stations requests
+- Bump pypdf to <7
 
 ## 0.112.0 - 2025-09-06