+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Feature
 - Make Mosmix and DMO a lot faster for multiple stations requests
 - Bump pypdf to <7
+- Make pypdf optional
 
 ## 0.112.0 - 2025-09-06